shift right to avoid possibility of int64 overflow in rw_distance calculation

+++ src/doom/r_defs.h
@@ -282,7 +282,7 @@ typedef struct
     sector_t*	frontsector;
     sector_t*	backsector;
     
-    int64_t	length; // [crispy] fix long wall wobble
+    uint32_t	length; // [crispy] fix long wall wobble
     angle_t	pangle; // [crispy] re-calculated angle used for rendering
 } seg_t;
